#151303085 INVESTIGATION: /contacts  API stubs POST should post a single contact
