+++ src/main/java/gov/ca/cwds/inject/ResourcesModule.java
@@ -1,14 +1,8 @@
 package gov.ca.cwds.inject;
 
-import com.google.inject.AbstractModule;
-import com.google.inject.Injector;
-import com.google.inject.Provides;
-import com.google.inject.name.Named;
-
 import gov.ca.cwds.rest.ApiConfiguration;
 import gov.ca.cwds.rest.SwaggerConfiguration;
 import gov.ca.cwds.rest.api.contact.DeliveredServiceDomain;
-import gov.ca.cwds.rest.api.domain.ContactRequestList;
 import gov.ca.cwds.rest.api.domain.IntakeLovEntry;
 import gov.ca.cwds.rest.api.domain.IntakeLovResponse;
 import gov.ca.cwds.rest.api.domain.StaffPerson;
@@ -33,6 +27,7 @@ import gov.ca.cwds.rest.api.domain.es.IndexQueryResponse;
 import gov.ca.cwds.rest.api.domain.investigation.HistoryOfInvolvement;
 import gov.ca.cwds.rest.api.domain.investigation.Investigation;
 import gov.ca.cwds.rest.api.domain.investigation.Relationship;
+import gov.ca.cwds.rest.api.domain.investigation.contact.ContactRequest;
 import gov.ca.cwds.rest.resources.AddressResource;
 import gov.ca.cwds.rest.resources.AddressValidationResource;
 import gov.ca.cwds.rest.resources.ApplicationResource;
@@ -64,8 +59,8 @@ import gov.ca.cwds.rest.resources.cms.ReferralClientResource;
 import gov.ca.cwds.rest.resources.cms.ReferralResource;
 import gov.ca.cwds.rest.resources.cms.ReporterResource;
 import gov.ca.cwds.rest.resources.cms.StaffPersonResource;
-import gov.ca.cwds.rest.resources.contact.ContactResource;
 import gov.ca.cwds.rest.resources.contact.DeliveredServiceResource;
+import gov.ca.cwds.rest.resources.investigation.ContactResource;
 import gov.ca.cwds.rest.resources.investigation.HistoryOfInvolvementResource;
 import gov.ca.cwds.rest.resources.investigation.RelationshipsResource;
 import gov.ca.cwds.rest.services.AddressService;
@@ -95,12 +90,17 @@ import gov.ca.cwds.rest.services.cms.ReferralService;
 import gov.ca.cwds.rest.services.cms.ReporterService;
 import gov.ca.cwds.rest.services.cms.StaffPersonService;
 import gov.ca.cwds.rest.services.cms.SystemCodeService;
-import gov.ca.cwds.rest.services.contact.ContactService;
 import gov.ca.cwds.rest.services.contact.DeliveredService;
 import gov.ca.cwds.rest.services.es.IndexQueryService;
+import gov.ca.cwds.rest.services.investigation.ContactService;
 import gov.ca.cwds.rest.services.investigation.HistoryOfInvolvementService;
 import gov.ca.cwds.rest.services.investigation.InvestigationService;
 import gov.ca.cwds.rest.services.investigation.RelationshipService;
+
+import com.google.inject.AbstractModule;
+import com.google.inject.Injector;
+import com.google.inject.Provides;
+import com.google.inject.name.Named;
 
 
 /**
@@ -193,8 +193,7 @@ public class ResourcesModule extends AbstractModule {
 
   @Provides
   @AllegationServieBackedResource
-  public TypedResourceDelegate<String, Allegation> allegationServiceBackedResource(
-      Injector injector) {
+  public TypedResourceDelegate<String, Allegation> allegationServiceBackedResource(Injector injector) {
     return new TypedServiceBackedResourceDelegate<>(injector.getInstance(AllegationService.class));
   }
 
@@ -208,8 +207,7 @@ public class ResourcesModule extends AbstractModule {
   @DrmsDocumentServiceBackedResource
   public TypedResourceDelegate<String, DrmsDocument> drmsDocumentServiceBackedResource(
       Injector injector) {
-    return new TypedServiceBackedResourceDelegate<>(
-        injector.getInstance(DrmsDocumentService.class));
+    return new TypedServiceBackedResourceDelegate<>(injector.getInstance(DrmsDocumentService.class));
   }
 
   @Provides
@@ -222,7 +220,7 @@ public class ResourcesModule extends AbstractModule {
 
   @Provides
   @ContactServiceBackedResource
-  public TypedResourceDelegate<String, ContactRequestList> contactServiceBackedResource(
+  public TypedResourceDelegate<String, ContactRequest> contactServiceBackedResource(
       Injector injector) {
     return new TypedServiceBackedResourceDelegate<>(injector.getInstance(ContactService.class));
   }
@@ -239,8 +237,8 @@ public class ResourcesModule extends AbstractModule {
   @ScreeningSummaryServiceBackedResource
   public TypedResourceDelegate<String, gov.ca.cwds.rest.api.domain.investigation.ScreeningSummary> screeningSummaryServiceBackedResource(
       Injector injector) {
-    return new TypedServiceBackedResourceDelegate<>(injector
-        .getInstance(gov.ca.cwds.rest.services.investigation.ScreeningSummaryService.class));
+    return new TypedServiceBackedResourceDelegate<>(
+        injector.getInstance(gov.ca.cwds.rest.services.investigation.ScreeningSummaryService.class));
   }
 
   @Provides
@@ -292,8 +290,7 @@ public class ResourcesModule extends AbstractModule {
   @Provides
   @ScreeningToReferralServiceBackedResource
   public ResourceDelegate screeningToReferralBackedResource(Injector injector) {
-    return new ServiceBackedResourceDelegate(
-        injector.getInstance(ScreeningToReferralService.class));
+    return new ServiceBackedResourceDelegate(injector.getInstance(ScreeningToReferralService.class));
   }
 
   @Provides
@@ -357,8 +354,7 @@ public class ResourcesModule extends AbstractModule {
 
   @Provides
   @AssignmentServiceBackedResource
-  public TypedResourceDelegate<String, Assignment> assignmentServiceBackedResource(
-      Injector injector) {
+  public TypedResourceDelegate<String, Assignment> assignmentServiceBackedResource(Injector injector) {
     return new TypedServiceBackedResourceDelegate<>(injector.getInstance(AssignmentService.class));
   }
 
@@ -401,18 +397,15 @@ public class ResourcesModule extends AbstractModule {
 
   @Provides
   @InvestigationServiceBackedResource
-  public TypedResourceDelegate<String, Investigation> investigationBackedResource(
-      Injector injector) {
+  public TypedResourceDelegate<String, Investigation> investigationBackedResource(Injector injector) {
     return new TypedServiceBackedResourceDelegate<>(
         injector.getInstance(InvestigationService.class));
   }
 
   @Provides
   @RelationshipsServiceBackedResource
-  public TypedResourceDelegate<String, Relationship> reslationshipBackedResource(
-      Injector injector) {
-    return new TypedServiceBackedResourceDelegate<>(
-        injector.getInstance(RelationshipService.class));
+  public TypedResourceDelegate<String, Relationship> reslationshipBackedResource(Injector injector) {
+    return new TypedServiceBackedResourceDelegate<>(injector.getInstance(RelationshipService.class));
 
   }
