+++ src/main/java/gov/ca/cwds/rest/api/domain/investigation/contact/ContactRequest.java
@@ -1,17 +1,14 @@
-package gov.ca.cwds.rest.api.domain;
+package gov.ca.cwds.rest.api.domain.investigation.contact;
 
-import static gov.ca.cwds.data.persistence.cms.CmsPersistentObject.CMS_ID_LEN;
-import gov.ca.cwds.data.persistence.contact.DeliveredServiceEntity;
 import gov.ca.cwds.rest.api.Request;
-import gov.ca.cwds.rest.api.Response;
+import gov.ca.cwds.rest.api.domain.DomainObject;
+import gov.ca.cwds.rest.api.domain.PostedIndividualDeliveredService;
+import gov.ca.cwds.rest.api.domain.SystemCodeCategoryId;
 import gov.ca.cwds.rest.validation.ValidSystemCodeId;
 import io.dropwizard.jackson.JsonSnakeCase;
 import io.dropwizard.validation.OneOf;
 import io.swagger.annotations.ApiModelProperty;
 
-import java.text.DateFormat;
-import java.text.SimpleDateFormat;
-import java.util.Date;
 import java.util.Set;
 
 import javax.validation.constraints.Size;
@@ -19,7 +16,6 @@ import javax.validation.constraints.Size;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.hibernate.validator.constraints.NotEmpty;
-import org.jadira.usertype.spi.utils.lang.StringUtils;
 
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonFormat;
@@ -27,38 +23,24 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
 /**
- * {@link DomainObject} representing a Contact
+ * {@link DomainObject} representing a Contact Request
  * 
  * @author CWDS API Team
  */
 @JsonSnakeCase
-@JsonPropertyOrder({"id", "lastUpdatedBy", "staffName", "startedAt", "endedAt", "purpose",
-    "communicationMethod", "status", "services", "location", "note", "people"})
-public class Contact extends ReportingDomain implements Request, Response {
+@JsonPropertyOrder({"startedAt", "endedAt", "purpose", "communicationMethod", "status", "services",
+    "location", "note", "people"})
+public class ContactRequest implements Request {
 
   /**
    * 
    */
   private static final long serialVersionUID = 1L;
 
-  @Size(max = CMS_ID_LEN)
-  // @NotNull
-  @JsonProperty("id")
-  @ApiModelProperty(required = true, readOnly = true, value = "", example = "ABC1234567")
-  private String id;
-
-  @JsonProperty("last_updated_by")
-  @ApiModelProperty(required = false, readOnly = false)
-  // , value = "primary contact staff person id")
-  private LastUpdatedBy lastUpdatedBy;
-
-
   @NotEmpty
   @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'")
-  // datetime
   @JsonProperty("started_at")
   @gov.ca.cwds.rest.validation.Date(format = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'", required = true)
-  // "2017-08-28T14:30:00.000Z"
   @ApiModelProperty(required = true, readOnly = false, value = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'",
       example = "2010-04-27T23:30:14.000Z")
   private String startedAt;
@@ -72,13 +54,9 @@ public class Contact extends ReportingDomain implements Request, Response {
   private String endedAt;
 
   @JsonProperty("purpose")
-  // "service_contact" delivered service serviceContactType
   @ValidSystemCodeId(required = false, category = SystemCodeCategoryId.CONTACT_TYPE)
-  @ApiModelProperty(
-      required = false,
-      readOnly = false,
-      value = "Delivered service contact type system code ID e.g)  -> 433 Conduct Client Evaluation ",
-      example = "433")
+  @ApiModelProperty(required = false, readOnly = false,
+      value = "Delivered service contact type system code ID e.g)  -> ", example = "433")
   private Integer purpose;
 
   @JsonProperty("communication_method")
@@ -97,12 +75,10 @@ public class Contact extends ReportingDomain implements Request, Response {
 
 
   @JsonProperty("services")
-  // need clarification
   @ApiModelProperty(required = false, readOnly = false)
   private Set<Integer> services;
 
   @JsonProperty("location")
-  // ("contact_location")
   @ValidSystemCodeId(required = false, category = SystemCodeCategoryId.CONTACT_LOCATION)
   @ApiModelProperty(required = false, readOnly = false,
       value = "Delivered service contact location type system code ID e.g) 415 -> CWS Office",
@@ -110,8 +86,6 @@ public class Contact extends ReportingDomain implements Request, Response {
   private Integer location;
 
   @JsonProperty("note")
-  // ("detail_text")
-  // a row in LONG_TEXT with content of contact_narrative
   @Size(max = 8000)
   @ApiModelProperty(required = false, readOnly = false, value = "", example = "detail text")
   private String note;
@@ -122,8 +96,6 @@ public class Contact extends ReportingDomain implements Request, Response {
   private Set<PostedIndividualDeliveredService> people;
 
   /**
-   * @param id id
-   * @param lastUpdatedBy last updated by staff
    * @param startedAt started at
    * @param endedAt ended at
    * @param purpose purpose
@@ -135,17 +107,12 @@ public class Contact extends ReportingDomain implements Request, Response {
    * @param people people
    */
   @JsonCreator
-  public Contact(@JsonProperty("id") String id,
-      @JsonProperty("last_updated_by") LastUpdatedBy lastUpdatedBy,
-      @JsonProperty("started_at") String startedAt, @JsonProperty("ended_at") String endedAt,
-      @JsonProperty("purpose") Integer purpose,
+  public ContactRequest(@JsonProperty("started_at") String startedAt,
+      @JsonProperty("ended_at") String endedAt, @JsonProperty("purpose") Integer purpose,
       @JsonProperty("communication_method") Integer communicationMethod,
       @JsonProperty("status") String status, @JsonProperty("services") Set<Integer> services,
       @JsonProperty("location") Integer location, @JsonProperty("note") String note,
       @JsonProperty("people") Set<PostedIndividualDeliveredService> people) {
-    super();
-    this.id = id;
-    this.lastUpdatedBy = lastUpdatedBy;
     this.startedAt = startedAt;
     this.endedAt = endedAt;
     this.purpose = purpose;
@@ -156,61 +123,6 @@ public class Contact extends ReportingDomain implements Request, Response {
     this.note = note;
     this.people = people;
   }
-
-  public Contact(DeliveredServiceEntity persistedDeliverdService, LastUpdatedBy lastUpdatedBy,
-      String note, Set<PostedIndividualDeliveredService> people) {
-    super();
-    this.id = persistedDeliverdService.getId();
-    this.lastUpdatedBy = lastUpdatedBy;
-    String startDate = DomainChef.cookDate(persistedDeliverdService.getStartDate());
-    if (StringUtils.isNotEmpty(startDate)) {
-      this.startedAt = startDate + "T" + cookTime(persistedDeliverdService.getStartTime()) + "Z";
-    }
-    String endDate = DomainChef.cookDate(persistedDeliverdService.getEndDate());
-    if (StringUtils.isNotEmpty(endDate)) {
-      this.endedAt = endDate + "T" + cookTime(persistedDeliverdService.getEndTime()) + "Z";
-    }
-    this.purpose = new Integer(persistedDeliverdService.getServiceContactType());
-    this.communicationMethod = new Integer(persistedDeliverdService.getCommunicationMethodType());
-    this.status = persistedDeliverdService.getStatusCode();
-    this.services = null;
-    this.location = new Integer(persistedDeliverdService.getContactLocationType());
-    this.note = note;
-    this.people = people;
-  }
-
-  public Contact() {
-    // default
-  }
-
-  /**
-   * @param date date to cook
-   * @return String in TIME_FORMAT
-   */
-  public static String cookTime(Date date) {
-    if (date != null) {
-      DateFormat df = new SimpleDateFormat("HH:mm:ss.SSS");
-      return df.format(date);
-    }
-    return "00:00:00.000";
-  }
-
-  /**
-   * @return the id
-   */
-  public String getId() {
-    return id;
-  }
-
-
-
-  /**
-   * @return the lastUpdatedBy
-   */
-  public LastUpdatedBy getLastUpdatedBy() {
-    return lastUpdatedBy;
-  }
-
 
 
   /**
